Bug 60656 - Emf image support in slideshows - use Rectangle2D instead of Dimension2D for image bounds - fix shearing transformation - fix rendering of font attributes (bold/italic/...) - emf+: needs its own object table and properties table - emf+: add linear gradient handler - emf+: handle brush data of pens - wmf/emf/emf+: position right aligned text correctly - emf+: use emf+ instead of emf records in dual-mode - emf+: handle region data and operations correctly - emf/+: map font weight to awt font weight correctly
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1870566 13f79535-47bb-0310-9956-ffa450edef68

+++ src/scratchpad/src/org/apache/poi/hwmf/record/HwmfFont.java
@@ -288,7 +288,7 @@ public class HwmfFont implements FontInfo, GenericRecord {
      * For all height comparisons, the font mapper SHOULD find the largest physical
      * font that does not exceed the requested size.
      */
-    protected int height;
+    protected double height;
 
     /**
      * A 16-bit signed integer that defines the average width, in logical units, of
@@ -433,7 +433,7 @@ public class HwmfFont implements FontInfo, GenericRecord {
         facename = "SansSerif";
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
